Use generic term 'test statistic'
Use 'test_stat' over 'qtilde' as to not assume the type of test being
performed. This should probably be something that is worked on across
the code base to harmonize languge and avoid language that makes the
user know information about how a test stat will be used:

c.f. pyhf.infer.qmu

Hail Jupyter